# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel API",
        "description": "API para gerenciamento de viagens",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/selection-process/search": {
            "get": {
                "tags": [
                    "Processos Seletivos"
                ],
                "summary": "Listar processos seletivos",
                "operationId": "53d99f33327feaf680957917e946ca30",
                "parameters": [
                    {
                        "name": "search_term",
                        "in": "query",
                        "description": "Termo para busca",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar por processos seletivos ativos",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de processos seletivos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/selection-process/{id}": {
            "get": {
                "tags": [
                    "Processos Seletivos"
                ],
                "summary": "Obter detalhes de um processo seletivo",
                "operationId": "de3bf92e45404c8df83cea29eed7919b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do processo seletivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do processo seletivo retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Processos Seletivos"
                ],
                "summary": "Deletar um processo seletivo",
                "operationId": "e382d44434c1f27301964b7f3f2eb9f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do processo seletivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Processo seletivo deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Processos Seletivos"
                ],
                "summary": "Atualizar um processo seletivo",
                "operationId": "bd85054fc6cd61a5f039c8cee4425435",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do processo seletivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título do processo seletivo",
                                        "type": "string"
                                    },
                                    "total_candidates": {
                                        "description": "Número total de candidatos",
                                        "type": "integer"
                                    },
                                    "available_vacancies": {
                                        "description": "Número de vagas disponíveis",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "vacancy_id": {
                                        "description": "ID da vaga associada",
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "description": "Status do processo seletivo (ativo ou inativo)",
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Processo seletivo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/selection-process/create": {
            "post": {
                "tags": [
                    "Processos Seletivos"
                ],
                "summary": "Criar um novo processo seletivo",
                "operationId": "6dbfe8f050f7db4cffc19cb0630c96f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título do processo seletivo",
                                        "type": "string"
                                    },
                                    "total_candidates": {
                                        "description": "Número total de candidatos",
                                        "type": "integer"
                                    },
                                    "available_vacancies": {
                                        "description": "Número de vagas disponíveis",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "vacancy_id": {
                                        "description": "ID da vaga associada",
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "description": "Status do processo seletivo (ativo ou inativo)",
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Processo seletivo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT obtido no login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}